fixes: added checking rows.Err, renamed accrual_system in accrual
